requirements.txt
# Investment Scouting & Diligence Engine - Dependencies

# Core data fetching
yfinance>=0.2.28
requests>=2.31.0

# AI/LLM
anthropic>=0.25.0

# Data processing
pandas>=2.0.0

# Web UI
streamlit>=1.28.0

# Google APIs (optional - for cloud export)
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-api-python-client>=2.95.0

# Environment management
python-dotenv>=1.0.0

# Optional: Excel export fallback
openpyxl>=3.1.0
